requirements.txt
streamlit
langchain
langchain_anthropic
langchain_community
boto3
anthropic
python-dotenv
openai
Pillow==8.3.2
langchain_openai